Upgradable factory and FraktalNFT airdrop receiver implementation(WIP)
