+++ contracts/PaymentSplitterUpgradeable.sol
@@ -2,7 +2,8 @@
 pragma solidity ^0.8.0;
 
 import "./IFraktalNFT.sol";
-import "./FraktalMarketV1_01.sol";
+import "./FraktalNFT.sol";
+import "./FraktalMarket.sol";
 import "@openzeppelin/contracts-upgradeable/utils/AddressUpgradeable.sol";
 import "@openzeppelin/contracts-upgradeable/utils/ContextUpgradeable.sol";
 import "@openzeppelin/contracts-upgradeable/proxy/utils/Initializable.sol";
@@ -50,8 +51,8 @@ contract PaymentSplitterUpgradeable is Initializable, ContextUpgradeable {
   {
     __PaymentSplitter_init(payees, shares_);
     tokenParent = _msgSender();
-    fraktionsIndex = IFraktalNFT(_msgSender()).getFraktionsIndex();
-    buyout = IFraktalNFT(_msgSender()).getStatus();
+    fraktionsIndex = FraktalNFT(_msgSender()).fraktionsIndex();
+    buyout = FraktalNFT(_msgSender()).sold();
     marketContract = _marketContract;
   }
 
@@ -149,7 +150,7 @@ contract PaymentSplitterUpgradeable is Initializable, ContextUpgradeable {
     _totalReleased = _totalReleased + payment;
 
     address payable marketPayable = payable(marketContract);
-    uint16 marketFee = FraktalMarketV1_1(marketPayable).fee();
+    uint16 marketFee = FraktalMarket(marketPayable).fee();
 
     uint256 forMarket = (payment * marketFee )/ 10000;
     uint256 forOperator = payment - forMarket;
